==> data_analysis/LICOR/LICOR_LOGS/03_explore_data.qmd <==
---
title: "LI-COR Data Exploration"
author: "Bryan Blue"
e-mail: 'bryanblue@arizona.edu'
date: "Last Generated: `r Sys.Date()`"
execute:
  echo: false
format: 
  html:
    code-fold: true
    fig-width: 8
    fig-height: 6
  pdf:
    fig-width: 7
    fig-height: 4
    echo: false

---
NOTE to self: Look into these for help with graph layout  
patchwork TODO !!!!  
wrap plot, axis collect

```{r setup, include=FALSE}
#| warning: false
#| echo: false
#| error: false

knitr::opts_chunk$set(echo = FALSE)
require(tidyverse)
library(plyr)
library(lubridate)
require(tidyr)
library(ggplot2)
library(here)
library(patchwork)
source(here("LICOR_LOGS_functions.R"))

# start in the current project directory
# this ensures relative paths will work
# independent of the location the script
here::here()
# https://www.regextester.com/ for regex checking

# constants to data paths
DATARAW <- "data_raw" 
DATACLEAN <- "data_clean"
DATAUSER <- "data_user"

```
*Always Restart R before running any script!*  
  
## NOTE: Run 02_load_clean_LICOR_logs.qmd first!  
  
Use "final_raw_data.csv" in the "data_user" folder.   
  
The data consists of the variables of interest found in the "select_logdata_fields.csv" file in the "data_user/input/" folder that were assembled in previous steps.
  
## Data Validation  
  
Many values in the LI-COR data contain invalid values such as any that are < 0. Any value less than 0 is converted to NA over the entire dataframe.  
  
The clock was off at the beginning of the experiment. The data are filtered for values of when the experiment started SysObs_time > 1668543540  
  
Only "Data_leaftype" values of "Treatment", "Control", and "Reference" are used. Any other user assigned values are ignored.   
Outliers: Many variables show distant outliers in graphs. Outlier routines based on IQR were tried.    
It did not work on the entire variable's data because it is not linear. Data at the tails were removed.  
It was also tried on a daily group. This did not work as there were many sets of data per day that only contained a couple values, or showed very high variance. This resulted in the complete drop of that day's values.  
Outliers were not removed for this data exploration.  
  
{{< pagebreak >}}

```{r daily_mean_data}
#| warning: false
#| echo: false
#| error: false

# calculate the daily means of a given column in sysobs
mean_plant_data <- function(finaldata, plantid, sysobs) {
  plotdf <- finaldata %>% filter(Data_plant_id == plantid) %>%
    filter(Data_leaftype == "Treatment" | 
             Data_leaftype == "Control" | 
             Data_leaftype == "Reference") %>%
    arrange(SysObs_date)
  
  plotdf <- plotdf %>%
    group_by(Filenames_filename, Data_leaftype, SysObs_date) %>%
    summarise_at((sysobs), list(gm = mean))
  
  return(plotdf)
}

plot_means <- function(plot_data, title="", xlabel="", ylabel=""){
  plot1 <- ggplot() +
    geom_point(data=plot_data, 
               aes(x = SysObs_date, y = gm, color = Data_leaftype)) +
    # geom_line(data=plot_data, 
    # aes(x = SysObs_date, y = gm, color = Data_leaftype)) +
        scale_color_manual(name='Leaf',
                       breaks=c('Treatment', 'Control', 'Reference'),
                       values=c('Treatment'='#F8766D', 'Control'='#619CFF', 'Reference'='#00BA38'))+
    ggtitle(title) +
    xlab(xlabel) +
    ylab(ylabel) +
    # scale_y_continuous(limits=c(0.0, 6.0))   +
    #     date_breaks = "10 sec",  expand = expansion(0)) +
    geom_smooth(se=FALSE, col='#F8766D', size=2) +
    theme(axis.text.x=element_text(angle=60, hjust=1))
  
  return(plot1)
}
```

```{r plot_box_plant}
#| warning: false
#| echo: false
#| error: false

plot_plant_line <- function(finaldata, plantid, fname, plantvar) {
  # TODO the filter for the leaftype needs removed, it should graph
  # every unique leaftype passed into it, data filtering before this call
  plant <- finaldata[finaldata$Filenames_filename == fname &
                       finaldata$Data_plant_id == plantid &
                       (finaldata$Data_leaftype == "Treatment" |
                          finaldata$Data_leaftype == "Control" |
                          finaldata$Data_leaftype == "Reference"), ]
  
  if (nrow(plant) > 0) {
      
    # display the variable name, not the internal name
      dispvarname <- str_split(plantvar, "_")[[1]][2]
      
       plot <- ggplot() +
      # geom_point(data = final_data, 
      geom_line(data=plant, 
                aes(x = SysObs_obs, y = get(plantvar), color = Data_leaftype)) +
      # geom_line(data = final_data, aes(x = SysObs_date, y = GasEx_gsw, color = "GasEx_gsw")) +
      ggtitle(paste("Level ", plantid, " LI-COR Readings ", fname)) +
      xlab("Observation Number") +
      ylab(dispvarname) +
    scale_color_manual(name='Leaf',
                       breaks=c('Treatment', 'Control', 'Reference'),
                       values=c('Treatment'='#F8766D', 'Control'='#619CFF', 'Reference'='#00BA38'))+
      # scale_y_continuous(limits=c(0.0, 0.2), breaks=c(0, 0.1, 0.2, 0.5))   +
      # scale_x_datetime(date_labels = ("%M:%S"),
      #     date_breaks = "10 sec",  expand = expansion(0)) +
      theme(axis.text.x=element_text(angle=60, hjust=1))
  }  
  return(plot)
}

variable_boxplot <- function(finaldata, lclevel, lcvariable) {
  
  elevation <- mean_plant_data(finaldata, lclevel, lcvariable)
  # display the variable name, not the internal name
  dispvarname <- str_split(lcvariable, "_")[[1]][2]
  
  p <- ggplot() +
    geom_boxplot(data = elevation, 
                 aes(x = Data_leaftype , y = gm, color = Data_leaftype), 
                 outlier.color = "#F8766D", 
                 outlier.size = 2) +
    # scale_y_continuous(limits=c(1.0, 2.8)) +#, breaks=c(0, 2.5, 3.0, 3.5)) +
    ggtitle(paste("LI-COR Observations for Elevation ", lclevel, " on ", dispvarname, sep = '')) +
    scale_color_manual(name='Leaf',
                       breaks=c('Treatment', 'Control', 'Reference'),
                       values=c('Treatment'='#F8766D', 'Control'='#619CFF', 'Reference'='#00BA38'))+
    xlab("Leaf Type") +
    ylab(dispvarname)
  
  return(p)
  
}

```


```{r remove_outliers}
#| warning: false
#| echo: false
#| error: false

# -----------------
# (https://www.reneshbedre.com/blog/find-outliers.html)

# look at all data for a given variable and look for outliers
# this does NOT work as the data is not linear
# values at the high or low ends can be removed
remove_outliers <- function(indata, elevation, leaftype, varname) {
  # elevation <- 3
  # varname <- "GasEx_A"
  # indata <- finaldata
  # 
  # leaftype <- "Reference"
  # leaftype <- "Control"
  # leaftype <- "Treatment"
  
  
  raw_data <- indata %>% filter(Data_plant_id == elevation 
                                & get(varname) > 0 
                                & Data_leaftype == leaftype)
  
  
  x <- raw_data %>% select(varname)
  # turn x into a vector
  x <- x[[varname]]
  
  
  # get values of Q1, Q3, and IQR
  # 8 => The resulting quantile estimates are approximately median-unbiased 
  # regardless of the distribution of x.
  sm <- summary(x, digits = 4)
  
  # get IQR
  thresh <- IQR(x)
  
  Tmin = sm[2] - (3 * thresh)
  Tmax = sm[5] + (3 * thresh)
  
  # find outlier
  newlist <- x[which(x < Tmin | x > Tmax)]
  newlist
  
  # print(paste("Removed: ", newlist, sep = '') 
  
  # remove outlier
  rmout <- x[which(x > Tmin & x < Tmax)]
  rmout
  
  # print(paste("OUTLIERS: ", cat(newlist), sep = '' ))
  
  # print(paste("RESULT: ", cat(rmout)))
  
  no_outliers <- raw_data[ raw_data[[varname]] %in% rmout, ]
  
  # raw_data <- filter(raw_data, (GasEx_A != newlist[1])) # does not retain NA
  
  
  return(no_outliers)
}

# Processing at a set of six readings does not work!
# Too few data points, and there are many invalid entries with NA values
remove_observation_outliers <- function(indata, elevation, leaftype, varname) {
  # testing data
  # indata <- finaldata
  # filename <- "2023-12-15-1126_logdata"
  # elevation <- 2
  # leaftype <- "Treatment"
  # varname <- "GasEx_A"
  # gr <- filename
  
  # init df for final output
  no_outliers <- data.frame()
  
  # break into observation groups and process at this level
  # not entire population of leaf type
  groups <- unique(indata$Filenames_filename)
  for (gr in groups) {
    
    # select all data from one log, for a given elevation and leaftype
    outlier_data <- filter(indata, (Filenames_filename == gr &
                                      Data_plant_id == elevation & 
                                      Data_leaftype == leaftype))
    
    x <- outlier_data[,varname]
    x <- x[[varname]]
    # get values of Q1, Q3, and IQR
    sm <- summary(x, digits = 8)
    
    # get IQR
    thresh <- IQR(x)
    
    Tmin = sm[2] - (1.5 * thresh) 
    Tmax = sm[5] + (1.5 * thresh) 
    
    # find outlier
    newlist <- x[which(x < Tmin | x > Tmax)]
    # newlist
    print(paste("Data Group: ", gr, " OUTLIERS: ", cat(newlist), sep = '' ))
    
    # remove outlier
    rmout <- x[which(x > Tmin & x < Tmax)]
    # rmout
    
    no_outliers <- rbind(no_outliers, outlier_data[ outlier_data[[varname]] %in% rmout, ])
    
  }
  
  return(no_outliers)
  
}

```
# Plots By Variable per Day
  
Every daily observation data set is plotted based on a variable of interest.  This variable can be easily changed to explore other data sets. 
```{r plot_var_per_day}
#| warning: false
#| echo: false
#| error: false

# t <- finaldata %>% filter(Data_leaftype == "Reference") %>% select(GasEx_A)

# TODO testing, hard code for now
# elevation <- 2
# filename <- "2023-12-15-1126_logdata"
varname <- "GasEx_A"


# remove outliers based on the ENTIRE SET of data
# outlier_dataT2 <- remove_outliers(finaldata, 2, "Treatment", varname)
# outlier_dataC2 <- remove_outliers(finaldata, 2, "Control", varname)
# outlier_dataR2 <- remove_outliers(finaldata, 2, "Reference", varname)
# outlier_data <- rbind(outlier_dataT2, outlier_dataC2, outlier_dataR2)
# 
# outlier_dataR3 <- remove_outliers(finaldata, 3, "Reference", varname)
# outlier_dataT3 <- remove_outliers(finaldata, 3, "Treatment", varname)
# outlier_dataC3 <- remove_outliers(finaldata, 3, "Control", varname)
# finaldata <- rbind(outlier_data, outlier_dataT3, outlier_dataC3, outlier_dataR3)

# remove outliers on an OBSERVATION SET of values 
# outlier_dataT2 <- remove_observation_outliers(finaldata, 2, "Treatment", varname)
# outlier_dataC2 <- remove_observation_outliers(finaldata, 2, "Control", varname)
# outlier_dataR2 <- remove_observation_outliers(finaldata, 2, "Reference", varname)
# outlier_data <- rbind(outlier_dataT2, outlier_dataC2, outlier_dataR2)
# 
# outlier_dataR3 <- remove_observation_outliers(finaldata, 3, "Reference", varname)
# outlier_dataT3 <- remove_observation_outliers(finaldata, 3, "Treatment", varname)
# outlier_dataC3 <- remove_observation_outliers(finaldata, 3, "Control", varname)
# finaldata <- rbind(outlier_data, outlier_dataT3, outlier_dataC3, outlier_dataR3)

# RAW DATA PLOTS
# process_data <- finaldata
# process_data <- no_zero_data

finaldata <- read_and_clean_data("final_raw_data.csv")


# finaldata <- process_data
for (FileName in unique(finaldata$Filenames_filename)) {
  # print(paste("Generate Plot: ", FileName, " PlantID: ", 
  # print(FileName)
  
  # plot one days readings as a line chart
  myplot <- plot_plant_line(finaldata, 
                            plantid = finaldata[finaldata$Filenames_filename==FileName,]$Data_plant_id, 
                            FileName, 
                            varname)
  print(myplot)
}

```
__Where is 2024-01-17-1143_logdata_lvl2, group1, 11:51:05 ref, 11:51:05 ref, 2,	Reference	1__

{{< pagebreak >}}
# Box-Plots of ALL Data
Raw data from all observations are included in these boxplots.  
  
Elevation 2 graphs are the most complete and accurate.   
Elevation 3 did not have accurate data collection until mid-February. Thermocouple movement caused misleading readings on some dates.  
  
Variables were selected to explore ALL of the data. These can easily have more added by changing the calls to the plots elevation and variable.
{{< pagebreak >}}

```{r boxplotss}
#| warning: false
#| echo: false
#| error: false


variable_boxplot(finaldata, "2", "GasEx_A") 
variable_boxplot(finaldata, "3", "GasEx_A")

variable_boxplot(finaldata, "2", "GasEx_gsw")
variable_boxplot(finaldata, "3", "GasEx_gsw")

variable_boxplot(finaldata, "2", "GasEx_VPDleaf")
variable_boxplot(finaldata, "3", "GasEx_VPDleaf")

variable_boxplot(finaldata, "2", "Meas_Tleaf")
variable_boxplot(finaldata, "3", "Meas_Tleaf")

```

```{r graphs_with_error_bars}
#| warning: false
#| echo: false
#| error: false

# testing error bars in plots

# ELEVATION constant is already applied above
# remove any leaf type that is not part of the experiment
# ELEVATION <- 3
# VARNAME <- "GasEx_gsw"
# graphit <- function(ELEVATION, VARNAME) {
#   
#   dispvarname <- str_split(VARNAME, "_")[[1]][2]
#   print(dispvarname)  
#   
#   df_summary <- finaldata %>% filter(Data_plant_id == ELEVATION) %>%
#     filter(Data_leaftype == "Treatment" | 
#              Data_leaftype == "Control" | 
#              Data_leaftype == "Reference")
#   
#   
#   # ----------- code works but does not split line on missing data
#   # mean values based on a group, and remove any na values
#   esummary <- plyr::ddply(df_summary, .(Filenames_filename, Data_leaftype, SysObs_date), summarise,
#                           mean = round(mean(get(VARNAME)), 2),
#                           sd = round(sd(get(VARNAME)), 2))
#   # esummary <- na.omit(esummary)
#   
#   
#   p1 <- ggplot(esummary, 
#                aes(x = SysObs_date, y = mean, color = Data_leaftype, 
#                    group(Data_leaftype, SysObs_date))) +
#     geom_errorbar(aes(ymin = mean - sd, ymax = mean + sd),
#                   position = position_dodge(0.9), width = 0.2) +
#     geom_line(linewidth = 0.3) +
#     geom_point(size = 2) +
#     scale_color_manual(name='Leaf',
#                        breaks=c('Treatment', 'Control', 'Reference'),
#                        values=c('Treatment'='#F8766D', 'Control'='#619CFF', 'Reference'='#00BA38'))+
#     theme(axis.text.x=element_text(angle=60, hjust=1)) + 
#     scale_x_date(date_breaks = "1 week", date_labels =  "%b %d") +
#     ggtitle(paste("Weekly Mean ", dispvarname, " for Elevation ", ELEVATION, sep = '')) +
#     xlab("Week of Observation") +
#     ylab(paste("mean ", dispvarname, " (µmol m⁻² s⁻¹)", sep = '')) 
# }

```

# Daily MEANS  
 
Line graphs for daily mean values are plotted with error bars.  
 
Breaks in lines indicate that there is no data at a point in between the ends of the lines. These are normally results of invalid data for that day, such as values < 0 that were converted to NAs.  
{{< pagebreak >}}
  

```{r run_error_graphs}
#| warning: false
#| echo: false
#| error: false

# TODO WHY DOES THIS NOT WORK!!! CAN NOT PASS VARIABLES TO FUNCTION
# GLOBAL VARIABLES WORK
grapherror <-function() {
  # ELEVATION <- e
  # VNAME <- o
  dispvarname <- str_split(VNAME, "_")[[1]][2]
  
  # remove any leaf type that is not part of the experiment
  df_summary <- finaldata %>% filter(Data_plant_id == ELEVATION) %>%
    filter(Data_leaftype == "Treatment" | 
             Data_leaftype == "Control" | 
             Data_leaftype == "Reference")
  
  
  # ----------- code works but does not split line on missing data
  # mean values based on a group, and remove any NA values
  esummary <- ddply(df_summary, .(Filenames_filename, Data_leaftype, SysObs_date), summarise,
                    # data is too variable for mean, use median
                    mean = round(median(get(VNAME)), 2),
                    # mean = round(mean(get(VNAME)), 2),
                    sd = round(sd(get(VNAME)), 2))
  # esummary <- na.omit(esummary)
  
  p1 <- ggplot(esummary, 
               aes(x = SysObs_date, y = mean, color = Data_leaftype, 
                   group(Data_leaftype, SysObs_date))) +
    geom_errorbar(aes(ymin = mean - sd, ymax = mean + sd),
                  position = position_dodge(0.9), width = 0.2) +
    geom_line(linewidth = 0.3) +
    geom_point(size = 2) +
    scale_color_manual(name='Leaf',
                       breaks=c('Treatment', 'Control', 'Reference'),
                       values=c('Treatment'='#F8766D', 'Control'='#619CFF', 'Reference'='#00BA38'))+
    theme(axis.text.x=element_text(angle=60, hjust=1)) + 
    scale_x_date(date_breaks = "1 week", date_labels =  "%b %d") +
    ggtitle(paste("Weekly Median ", dispvarname, " for Elevation ", ELEVATION, sep = '')) +
    xlab("Week of Observation") +
    ylab(paste("median ", dispvarname, " (µmol m⁻² s⁻¹)", sep = '')) 
  
  return(p1)
  
}

# TODO WHY DOES THIS NOT WORK!!! CAN NOT PASS VARIABLES TO FUNCTION
# GLOBAL VARIABLES WORK
ELEVATION <- 2
VNAME <- "GasEx_A"
grapherror()
ELEVATION <- 3
VNAME <-"GasEx_A"
grapherror()

ELEVATION <- 2
VNAME <- "GasEx_gsw"
grapherror()
ELEVATION <- 3
VNAME <-"GasEx_gsw"
grapherror()

ELEVATION <- 2
VNAME <- "GasEx_VPDleaf"
grapherror()
ELEVATION <- 3
VNAME <-"GasEx_VPDleaf"
grapherror()

ELEVATION <- 2
VNAME <- "Meas_Tleaf"
grapherror()
ELEVATION <- 3
VNAME <-"Meas_Tleaf"
grapherror()


```




```{r test_error_bars2}
#| warning: false
#| echo: false
#| error: false
#| 
# # testing error bars in plots
# 
# ELEVATION <- 2
# VARNAME <- "GasEx_gsw"
# dispvarname <- str_split(VARNAME, "_")[[1]][2]
# 
# # remove any leaf type that is not part of the experiment
# df_summary <- finaldata %>% filter(Data_plant_id == ELEVATION) %>%
#   filter(Data_leaftype == "Treatment" | 
#            Data_leaftype == "Control" | 
#            Data_leaftype == "Reference")
# 
# 
# # ----------- code works but does not split line on missing data
# # mean values based on a group, and remove any NA values
# esummary <- ddply(df_summary, .(Filenames_filename, Data_leaftype, SysObs_date), summarise,
#                   mean = round(mean(get(VARNAME)), 2),
#                   sd = round(sd(get(VARNAME)), 2))
# # esummary <- na.omit(esummary)
# 
# p1 <- ggplot(esummary, 
#              aes(x = SysObs_date, y = mean, color = Data_leaftype, 
#                  group(Data_leaftype, SysObs_date))) +
#   geom_errorbar(aes(ymin = mean - sd, ymax = mean + sd),
#                 position = position_dodge(0.9), width = 0.2) +
#   geom_line(linewidth = 0.3) +
#   geom_point(size = 2) +
#   scale_color_manual(name='Leaf',
#                      breaks=c('Treatment', 'Control', 'Reference'),
#                      values=c('Treatment'='#F8766D', 'Control'='#619CFF', 'Reference'='#00BA38'))+
#   theme(axis.text.x=element_text(angle=60, hjust=1)) + 
#   scale_x_date(date_breaks = "1 week", date_labels =  "%b %d") +
#   ggtitle(paste("Weekly Mean ", dispvarname, " for Elevation ", ELEVATION, sep = '')) +
#   xlab("Week of Observation") +
#   ylab(paste("mean ", dispvarname, " (µmol m⁻² s⁻¹)", sep = '')) 
# 
# p1
# 
# p2 <- ggplot(esummary, aes(x=SysObs_date, y=mean, color = Data_leaftype, group(Data_leaftype, SysObs_date))) +
#   geom_errorbar(aes(ymin = mean - sd, ymax = mean + sd), 
#                 position = position_dodge(0.9), width = 0.2) +
#   # geom_line(size= 0.2) +
#   geom_point(size = 2) +
#   scale_color_manual(name='Leaf',
#                      breaks=c('Treatment', 'Control', 'Reference'),
#                      values=c('Treatment'='#F8766D', 'Control'='#619CFF', 'Reference'='#00BA38'))+
#   theme(axis.text.x=element_text(angle=60, hjust=1)) + 
#   scale_x_date(date_breaks = "1 week", date_labels =  "%b %d") +
#   ggtitle(paste("Weekly Mean ", dispvarname, " for Elevation ", ELEVATION, sep = '')) +
#   xlab("Week of Observation") +
#   ylab(paste("mean ", dispvarname, " (µmol m⁻² s⁻¹)", sep = '')) 
# 
# p2
# 
# p1/p2

```


```{r test_error_bars3}
#| warning: false
#| echo: false
#| error: false

# # testing error bars in plots
# 
# # ELEVATION constant is already applied above
# # remove any leaf type that is not part of the experiment
# ELEVATION <- 3
# VARNAME <- "GasEx_A"
# df_summary <- finaldata %>% filter(Data_plant_id == ELEVATION) %>%
#   filter(Data_leaftype == "Treatment" | 
#            Data_leaftype == "Control" | 
#            Data_leaftype == "Reference")
# 
# 
# # ----------- code works but does not split line on missing data
# # mean values based on a group, and remove any na values
# esummary <- plyr::ddply(df_summary, .(Filenames_filename, Data_leaftype, SysObs_date), summarise,
#                         mean = round(mean(get(VARNAME)), 2),
#                         sd = round(sd(get(VARNAME)), 2))
# # esummary <- na.omit(esummary)
# 
# 
# p1 <- ggplot(esummary, 
#              aes(x = SysObs_date, y = mean, color = Data_leaftype, 
#                  group(Data_leaftype, SysObs_date))) +
#   geom_errorbar(aes(ymin = mean - sd, ymax = mean + sd),
#                 position = position_dodge(0.9), width = 0.2) +
#   geom_line(linewidth = 0.3) +
#   geom_point(size = 2) +
#   scale_color_manual(name='Leaf',
#                      breaks=c('Treatment', 'Control', 'Reference'),
#                      values=c('Treatment'='#F8766D', 'Control'='#619CFF', 'Reference'='#00BA38'))+
#   theme(axis.text.x=element_text(angle=60, hjust=1)) + 
#   scale_x_date(date_breaks = "1 week", date_labels =  "%b %d") +
#   ggtitle(paste("Weekly Mean ", dispvarname, " for Elevation ", ELEVATION, sep = '')) +
#   xlab("Week of Observation") +
#   ylab(paste("mean ", dispvarname, " (µmol m⁻² s⁻¹)", sep = '')) 
# 
# p1
# 
# p2 <- ggplot(esummary, aes(x=SysObs_date, y=mean, color = Data_leaftype, group(Data_leaftype, SysObs_date))) +
#   geom_errorbar(aes(ymin = mean - sd, ymax = mean + sd), 
#                 position = position_dodge(0.9), width = 0.2) +
#   # geom_line(size= 0.2) +
#   geom_point(size = 2) +
#   scale_color_manual(name='Leaf',
#                      breaks=c('Treatment', 'Control', 'Reference'),
#                      values=c('Treatment'='#F8766D', 'Control'='#619CFF', 'Reference'='#00BA38'))+
#   theme(axis.text.x=element_text(angle=60, hjust=1)) + 
#   scale_x_date(date_breaks = "1 week", date_labels =  "%b %d") +
#   ggtitle(paste("Weekly Mean ", dispvarname, " for Elevation ", ELEVATION, sep = '')) +
#   xlab("Week of Observation") +
#   ylab(paste("mean ", dispvarname, " (µmol m⁻² s⁻¹)", sep = '')) 
# p2 
# 
# p1/p2

```
